finsihed simple interaction
added a way to move around the screen by clicking and dragging and a way
to zoom in and out.
It is all very basic, but we will keep it like this and use the zoom
variable to decide what labels to show and what not to show depending on
their importance.