simple label overlap fix
when labels overlap, the least important one is not drawn, this way the
important labels are visible from a far and when zoomed in the lesser
important labels are visible.
Also fixed a seemingly small bug in positioning, for this an additional
scaling is introduces after the initial placement. This scales the
positioning down untill it fits in the bounding polygon.
Also there is a movement limit on the normal nodes, this stops the
movement after a certain amount of iterations.