colorbrewer colours
added the colorbrewer colours. Instead of getting the json file and
parsing it and requesting the amount of colours we just hardcoded 12
colours. This is because if you requested colours, you would get it from
the same dataset of the 12 colours. By hardcoding 12 colours we save a
lot of trouble. Also we removed all the iterative functions to
immediatly give the end result